parse the code XML natively in node, and finish up displaying section body content
xmltodict destroys the document order of nodes with different tags, but <text> and <level> nodes are interleaved.
This necessitated a different approach to parsing the code files. I've rewritten it using elementtree, which seems
to be also the only XML parser for node that properly handles mixed-content elements like we have in <text> when
there is text formatting.